fix: changed u64 function taking a u64 to a u64

+++ libs/core/src/core/cint.h
@@ -54,7 +54,7 @@ namespace euphoria::core
 
 
     constexpr float
-    c_u64_to_float(u8 i)
+    c_u64_to_float(u64 i)
     {
         return static_cast<float>(i);
     }
